bug fixed

+++ src/services/user.services.ts
@@ -1,18 +1,26 @@
 import { IUser } from '../interfaces/user.Interface';
 import User from '../model/user.model';
+import bcrypt from 'bcrypt';
 
 const createUser = async (userData: IUser): Promise<IUser | null> => {
-  const result = await User.create(userData);
+  const existingUser = await User.findOne({ userId: String(userData.userId) });
+  if (existingUser) {
+    throw new Error('User already exists!!!!!!!!!!!!!!!!!!!!');
+  }
+  const hashedPassword = await bcrypt.hash(userData.password, 10);
+  const userWithHashedPassword = {
+    ...userData,
+    password: hashedPassword,
+  };
 
-  // Include _id as userId if that's the generated MongoDB ObjectId
+  const result = await User.create(userWithHashedPassword);
+
   const { _id: userId, password, ...restData } = result.toObject();
 
-  // Define a partial IUser without the 'password' property
   const partialUser: Partial<IUser> = {
     ...restData,
   };
 
-  // Type assertion to IUser, assuming password will be available in IUser
   const userWithoutPassword = partialUser as IUser;
 
   return userWithoutPassword;
@@ -26,6 +34,10 @@ const getAllUsers = async (): Promise<IUser[]> => {
 };
 
 const getSingleUserById = async (userId: number): Promise<IUser | null> => {
+  const existingUser = await User.findOne({ userId: String(userId) });
+  if (!existingUser) {
+    throw new Error('User not found!!!');
+  }
   const result = await User.findOne({ userId }).select('-password');
   return result;
 };
@@ -34,24 +46,22 @@ const updateUserById = async (
   userId: number,
   userData: IUser,
 ): Promise<IUser | null> => {
-  const result = await User.findOne({ userId });
-  if (result) {
-    result.set(userData);
-    await result.save();
-    const { _id, password, ...restData } = result.toObject();
-
-    // Define a partial IUser without the 'password' property
-    const partialUser: Partial<IUser> = {
-      ...restData,
-    };
-    const userWithoutPassword = partialUser as IUser;
-
-    return userWithoutPassword;
+  const existingUser = await User.findOne({ userId: String(userId) });
+  if (!existingUser) {
+    throw new Error('User not found!!!!!!!!!!!!!!!!!!!!');
   }
-  return null;
+
+  const result = await User.findOneAndUpdate({ userId }, userData, {
+    new: true,
+  }).select('-password');
+  return result;
 };
 
 const deleteUserById = async (userId: number): Promise<IUser | null> => {
+  const existingUser = await User.findOne({ userId: String(userId) });
+  if (!existingUser) {
+    throw new Error('User not found!!!');
+  }
   const result = await User.findOneAndDelete({ userId: userId });
   return result;
 };
